update and destroy policies added against job controller

+++ app/Http/Controllers/JobController.php
@@ -8,9 +8,14 @@ use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Support\Facades\Storage;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
+
 
 class JobController extends Controller
 {
+
+    use AuthorizesRequests;
+
     /**
      * Display a listing of the resource.
      */
@@ -86,6 +91,8 @@ class JobController extends Controller
      */
     public function edit(Job $job): View
     {
+        $this->authorize('update', $job);
+
         return view('jobs.edit')->with('job', $job);
     }
 
@@ -94,6 +101,10 @@ class JobController extends Controller
      */
     public function update(Request $request, Job $job): RedirectResponse
     {
+
+        # Check if user is authorized
+        $this->authorize('update', $job);
+
         $validatedData = $request->validate([
             'title' => 'required|string|max:255',
             'description' => 'required|string',
@@ -137,6 +148,9 @@ class JobController extends Controller
      */
     public function destroy(Job $job): RedirectResponse
     {
+        # Check if user is authorized
+        $this->authorize('delete', $job);
+
         // Delete logo file if exists
         if ($job->company_logo) {
             Storage::delete('public/' . basename($job->company_logo));
